refactor analyze file data structure
