+++ .eslintignore
@@ -4,5 +4,5 @@ pnpm-lock.yaml
 package-lock.json
 *.md
 // For the moment ignore these test files
-/tests/mocks/**
+/**/mocks/**
 /reports/**
